Move the remaining whitelabel plugin slots into the module
PLUGIN_WHITELABEL, PLUGIN_LANDING_PAGE, PLUGIN_HOMEPAGE_SETTING and
PLUGIN_LOGO_ICON_COMPONENTS are only ever written by the whitelabel plugin,
so they join PLUGIN_SELECTORS in metabase/whitelabel. metabase/plugins no
longer resets whitelabel's slots; whitelabel exports its own reinitialize.
