+++ enterprise/frontend/src/metabase-enterprise/whitelabel/index.ts
@@ -1,14 +1,14 @@
 import { t } from "ttag";
 
+import { PLUGIN_APP_INIT_FUNCTIONS } from "metabase/plugins";
+import MetabaseSettings from "metabase/utils/settings";
 import {
-  PLUGIN_APP_INIT_FUNCTIONS,
   PLUGIN_HOMEPAGE_SETTING,
   PLUGIN_LANDING_PAGE,
   PLUGIN_LOGO_ICON_COMPONENTS,
+  PLUGIN_SELECTORS,
   PLUGIN_WHITELABEL,
-} from "metabase/plugins";
-import MetabaseSettings from "metabase/utils/settings";
-import { PLUGIN_SELECTORS } from "metabase/whitelabel";
+} from "metabase/whitelabel";
 import { hasPremiumFeature } from "metabase-enterprise/settings";
 import {
   getApplicationName,
