+++ frontend/src/metabase/admin/settings/components/SettingsPages/AppearanceSettingsPage.tsx
@@ -1,6 +1,6 @@
 import { UpsellWhitelabel } from "metabase/admin/upsells";
 import { useHasTokenFeature } from "metabase/common/hooks";
-import { PLUGIN_WHITELABEL } from "metabase/plugins";
+import { PLUGIN_WHITELABEL } from "metabase/whitelabel";
 
 export function AppearanceSettingsPage({
   tab,
